Imgsearchengine/requirements.txt
alembic==1.0.8
analyzer==0.1.1
analyzerdam==0.1.2
analyzerstrategies==0.1.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.3.2
dicttoxml==1.7.4
Django==2.1.5
django-haystack==2.8.1
django-import-data==0.4.2
djangorestframework==3.9.2
googlefinance==0.7
graphviz==0.10.1
idna==2.8
igraph==0.1.11
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jgraph==0.2.1
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.3.2
Mako==1.0.8
MarkupSafe==1.1.1
matplotlib==3.0.3
mox==0.5.3
nbformat==4.4.0
networkx==2.2
numpy==1.16.2
pandas==0.24.2
parso==0.4.0
pbr==1.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.7.1
prompt-toolkit==2.0.9
public==2018.11.20
py-moneyed==0.8.0
Pygments==2.3.1
pyparsing==2.3.1
pyrsistent==0.14.11
pystock==0.1.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
query==0.1.4
requests==2.21.0
retrying==1.3.3
scikit-learn==0.20.3
scipy==1.2.1
self==2018.11.20
six==1.12.0
sklearn==0.0
soupsieve==1.9
SQLAlchemy==1.3.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Whoosh==2.7.4
xml2json==1.1
xmldataset==1.0.1
xmltodict==0.12.0
